+++ internal/thelma/charts/changedfiles/changedfiles.go
@@ -1,4 +1,4 @@
-// Package filetrigger maps a list of updated files in the terra-helmfile repo to a list of charts that need to be published.
+// Package changedfiles maps a list of updated files in the terra-helmfile repo to a list of charts that need to be published.
 //
 // The logic is as follows:
 // * Changes to charts/<chartname>/* or values/(app|cluster)/<chartname>* will trigger a publish of the affected chart
@@ -6,7 +6,7 @@
 // * Changes to values/cluster/global* will trigger a publish of all cluster release charts
 // * Changes to helmfile.yaml will trigger a publish/render of all charts that have at least one chart release
 // * Finally, any charts that are transitive dependencies of charts in the above list will be published as well
-package filetrigger
+package changedfiles
 
 import (
 	"bufio"
