+++ internal/thelma/charts/changedfiles/testdata/charts/agora/Chart.yaml
