+++ internal/thelma/charts/changedfiles/testdata/charts/foundation/Chart.yaml
