+++ internal/thelma/charts/changedfiles/testdata/charts/mysql/Chart.yaml
