+++ internal/thelma/charts/changedfiles/testdata/charts/postgres/Chart.yaml
