+++ internal/thelma/charts/changedfiles/testdata/charts/sam/Chart.yaml
