+++ internal/thelma/charts/changedfiles/testdata/charts/secrets-manager/Chart.yaml
