+++ internal/thelma/charts/changedfiles/testdata/charts/workspacemanager/Chart.yaml
